+++ UnderAutomation.UniversalRobots.ObfuscatedSources/Common/RobotModelsExtended.cs
@@ -1,0 +1,48 @@
+//              WARNING
+// This file is an empty shell containing only public C# items.
+// The internal code is hidden; to access it, you need to obtain a Source licence of the library.
+
+
+namespace Common {
+	/// <summary>
+	/// Model of a UR robot (including e-Series and extended payload models)
+	/// </summary>
+	public enum RobotModelsExtended {
+
+
+		UR3e = 0,
+
+
+		UR5e = 1,
+
+
+		UR7e = 2,
+
+
+		UR10e = 3,
+
+
+		UR12e = 4,
+
+
+		UR16e = 5,
+
+
+		UR15 = 6,
+
+
+		UR20 = 7,
+
+
+		UR30 = 8,
+
+
+		UR3 = 9,
+
+
+		UR5 = 10,
+
+
+		UR10 = 11,
+	}
+}