+++ UnderAutomation.UniversalRobots.ObfuscatedSources/Kinematics/Ur7eDhParameters.cs
@@ -1,0 +1,17 @@
+//              WARNING
+// This file is an empty shell containing only public C# items.
+// The internal code is hidden; to access it, you need to obtain a Source licence of the library.
+
+using Common;
+
+namespace Kinematics {
+
+	public class Ur7eDhParameters : Ur5eDhParameters, IUrDhParameters {
+
+
+		public Ur7eDhParameters()
+		{
+			// Source is hidden, a Source licence is needed to access internal code...
+		}
+	}
+}