+++ UnderAutomation.UniversalRobots.ObfuscatedSources/PrimaryInterface/ConfigurationDataPackageEventArgs.cs
@@ -8,7 +8,7 @@ namespace PrimaryInterface {
 	/// <summary>
 	/// Joint configuration
 	/// </summary>
-	public class ConfigurationDataPackageEventArgs : PackageEventArgs {
+	public class ConfigurationDataPackageEventArgs : PackageEventArgs, IUrDhParameters {
 
 
 		public ConfigurationDataPackageEventArgs()
@@ -105,5 +105,35 @@ public JointConfiguration Wrist2 { get; set; }
 		/// </summary>
 		[PackageDescription("Wrist3 (Tool) joint configuration")]
 public JointConfiguration Wrist3 { get; set; }
+
+		/// <summary>
+		/// DH parameter a2 (Shoulder.DHa)
+		/// </summary>
+		public double A2 { get; }
+
+		/// <summary>
+		/// DH parameter a3 (Elbow.DHa)
+		/// </summary>
+		public double A3 { get; }
+
+		/// <summary>
+		/// DH parameter d1 (Base.DHd)
+		/// </summary>
+		public double D1 { get; }
+
+		/// <summary>
+		/// DH parameter d4 (Wrist1.DHd)
+		/// </summary>
+		public double D4 { get; }
+
+		/// <summary>
+		/// DH parameter d5 (Wrist2.DHd)
+		/// </summary>
+		public double D5 { get; }
+
+		/// <summary>
+		/// DH parameter d6 (Wrist3.DHd)
+		/// </summary>
+		public double D6 { get; }
 	}
 }